# src/infrastructure/api/controllers/book/dto/post-book.dto.ts
import { z } from 'zod';
import { BookOutputDto } from './book.dto';

export const PostBookInputDto = z.object({
  title: z.string().min(1),
  summary: z.string().min(5),
  author: z.string().min(1),
  totalPages: z.number().min(1),
});
export type PostBookInputDto = ReturnType<typeof PostBookInputDto.parse>;

export const PostBookOutputDto = BookOutputDto;
export type PostBookOutputDtoBookOutputDto = ReturnType<
  typeof PostBookOutputDto.parse
>;